6a amd64: fix zaddr() long-vs-int32, fifth instance of this bug class
assemblers/6a/lex.c's zaddr() (the assembler-side object-file writer)
had the exact same bug already fixed in compilers/6c/swt.c's zaddr()
(the compiler-side writer): `long l` is 64 bits on our hosts, so
`l = a->offset; if((vlong)l != a->offset)` never differs and the T_64
object-file flag was never set, silently truncating any hand-assembled
64-bit immediate down to its low 32 bits -- e.g. "MOVQ
$0x7FF0000000000001,CX" assembled to the wrong 32-bit-immediate opcode
form (0xc7), encoding just $1 instead of the full value.

Found via a systematic grep for the `(vlong)l == /!=` truncation-check
idiom across assemblers/6a, compilers/6c and linkers/6l (prompted by
recalling the /tmp/diff6a /tmp/diff6c /tmp/diff6l cross-repository
comparisons from earlier this session) -- this is the fifth instance
of the same long-vs-int32-on-a-64-bit-host bug class this session,
after Ieee.l/.h, 6c's zaddr(), linkers/6l/span.c's oclass(), and
compilers/6c/cgen.c's immconst(). Go-era's own 6a (src/cmd/6a/lex.c
zaddr()) already uses int32 here.

Adds tests/s/variants/const_amd64.s: a minimal hand-assembled,
qemu-runnable regression test (verified it fails -- with a wrong
truncated value written to memory -- against the pre-fix assembler,
passes after).

+++ assemblers/6a/lex.c
@@ -1066,7 +1066,14 @@ zname(char *n, int t, int s)
 void
 zaddr(Gen *a, int s)
 {
-	long l;
+	/* claude: was `long l`. Same bug as compilers/6c/swt.c's zaddr():
+	 * on our 64-bit host `long` is 64 bits, so `l = a->offset;
+	 * if((vlong)l != a->offset)` never differs and T_64 was never set,
+	 * silently truncating any hand-assembled 64-bit immediate (e.g.
+	 * MOVQ $0x7FF0000000000001,CX) down to its low 32 bits. Go-era's
+	 * own 6a (src/cmd/6a/lex.c zaddr()) already uses int32 here.
+	 */
+	int32 l;
 	int i, t;
 	char *n;
 	Ieee e;
